+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [7.1.0-rc.1] - 2026-06-15
 
 ### Changed
 
@@ -1195,7 +1195,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
-[Unreleased]: https://https://github.com/internetguru/laravel-user/compare/staging...dev
+[7.1.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v7.0.2
 [7.0.2]: https://https://github.com/internetguru/laravel-user/compare/v7.0.1...v7.0.2
 [7.0.1]: https://https://github.com/internetguru/laravel-user/compare/v7.0.0...v7.0.1
 [7.0.0]: https://https://github.com/internetguru/laravel-user/compare/v6.7.2...v7.0.0